requirements.txt
aiogram
torch
torchvision
matplotlib
tensorflow
tensorflow-hub
scikit-learn
pandas
emoji
opencv-python
pyyaml